Add initial implementation of bootstrap.sh
This script is designed to be run on a vanilla virtual machine or even
previously configured machines to make sure they're ready to begin
building the liquorix package.

Although this script may take care of all the prerequisites, it's highly
recommended to set up a snapshot that has run this script at least once.
Failing to do so will incur a larger and more variable amount of
initialization time before your VM or direct hardware is ready to build.