Switch from full to voluntary kernel preemption + config refresh
In an effort to increase stability among newer laptops that focus on
long battery life, suspend times that last a month, and the importance
of mobile computing where it's common to suspend and resume multiple
times a day, the preemption configuration has been changed from a black
list to white list.

With full kernel preemption, critical sections are black listed
from being preemptable.  However, on newer hardware, it's not easy to
blacklist all critical sections until a bug has been reported.

On the other hand, with voluntary kernel preemption, well known sections
of the kernel that are preemptable are marked as such.  Switching to
this configuration should significantly increase stability on systems
that use proprietary or out-of-tree kernel modules, be it binary blobs
or not.

Although full kernel preemption does serve a significant user base of
Liquorix, this configuration is better served with an automated build
service that applies a configuration patch setting preemption from
voluntary to full, and then building multiple varieties for different
chipsets.