Switch back from MuQSS to CFS
Due to multiple issues I've been trying to correct on MuQSS, switching
Liquorix back to CFS from MuQSS.

Of them, MuQSS attempts to use all the CPUs for small tasks, causing the
load on every CPU to be very low for 100% tasks.  This confuses ondemand
and schedutil governors, creating very strange behavior when trying to
tune for power consumption.

Will revisit MuQSS once the governor problems for both ondemand and
schedutil are corrected.