Disable RCU_NOCB_CPU
This option was added to the kernel to save power on asymmetric
processors (like those found on phones with big.LITTLE).  This feature
also helps if you want the scheduler to determine when RCU can occur.

Unfortunately, this comes with a userspace <-> kernelspace overhead
which you can see visibly when monitoring CPU usage.

Disable since we won't see the full benefits unless we enable full
dynticks / adaptive ticks.