Enable TRANSPARENT_HUGEPAGE and TRANSPARENT_HUGEPAGE_ALWAYS
This option was originally disabled because of a benchmark online that
shows poor performance on a specific database.  Unfortunately, the
audience that's use Liquorix is putting this on their desktops and
laptops, not their servers.

Also, since changing to blk-mq, we've ruled out transparent hugepages as
the source of stutters and hangs during IO intensive operations.