Use BFQ-SQ for legacy block by default, disable blk-mq by default
Kernel 4.12 has a regression that causes the kernel to loop infinitely
when changing the power state of a device.  A good example is a laptop
that is unplugged or undocked, CPU time will hit 100% on a single kernel
thread forever until you plug the system back in.  This also prevents
the system from suspending as well.

For now, lets use BFQ-SQ on the legacy block subsystem until suspend
issues are resolved.