bootstrap.sh: Don't require root
In fact, if the user is root, exit unsuccessfully.  Running builds as
root breaks the local user's configurations and over-complicates the
build process.

If the build user requires root privileges, the owner of the system
needs to configure the build user to not require a password when running
sudo, or provide exceptions for the specific commands it will be using.