Restore debian nomenclature of initramfs-fallback
Upon further research on initramfs in Debian, both initramfs-tools and
dracut provide the package name "linux-initramfs-tool".  For
compatibility, initramfs-tools is mentioned explicitely while the
provided package name is mentioned as a fallback.

Since this is generic, lets just restore the initramfs generator
configuration back to how it was, more so that the diffs between related
kernel packages aren't different where not necessary.