delete-ppa-packages.py: Be smarter at removing binaries / sources
By definition, if a source is superseded, it's also not deleted yet.
Don't check the removal date since just the fact the source package
returns as superseded means it's not deleted yet.

When processing the binaries, don't check the removal date, check the
status.  Same as the source, if it's not 'Deleted', attempt to delete it
if it's currently 'Superseded'.