Correct EFI configuration, disable google firmware
Per a comment in the linux-lqx AUR, the EFI configuration causes errors
to print during boot-up.

In the Arch wiki for UEFI Variables [1], it's recommended that
CONFIG_EFI_VARS be disabled, and CONFIG_EFIVARS_FS be compiled
statically.

As collateral, google firmware was disabled since it had a hard
dependency on CONFIG_EFI_VARS, but if you read the description, it's for
proprietary hardware that consumers don't have access to anyway.

[1] https://wiki.archlinux.org/index.php/Unified_Extensible_Firmware_Interface#UEFI_Variables