env.sh: Use 'version_package' output as input for 'version_kernel'
This is necessary to prevent grep from matching multiple version strings
in the debian changelog if a nmu version is incremented.  By using the
version_package output as the input for version_kernel, we've already
reduced the data grep is matching with and prevent us from matching the
build version as well.