Restore 1000hz tick configuration
After switching to 250hz, Terence on the linux-lqx AUR comments
mentioned his audio configuration with 1.3ms of latency is getting 3-4x
more underruns than with the 1000hz config.

This prompted me to measure the system overhead of both 250hz and 1000z.
In powertop, on a 5960x running 1.5ghz while idle, the tick consumes
2-3ms.  With 1000hz, this quadruples to about 8ms while idle, but drops
to 5ms once the core frequency ramps up to baseclock + turbo.  In this
case, 8ms is under 1% of system overhead.

Lets keep Liquorix at 1000hz if it means audio hardware is less likely
to prompt underruns under low latency configurations.