Switch to shared runqueues for multi-core siblings
Regardless of whether we have a RQ for every processor core, or a shared
RQ for every core on a processor, this doesn't really keep running tasks
to the same core for higher throughput.  MuQSS will still try to steal a
task from another runqueue if the next eligible task on any of the other
RQ's is worse than its own next eligible task.

Lets keep the shared runqueue on the default, since its most likely to
be the most tested and dogfed configuration.