Revert "Switch to shared runqueues for multi-core siblings"
This reverts commit 83d03a668fc6176f83c96d3e84e61f87dea6b95a.

In a strange turn of events, shared runqueues may be causing strange
boot stall problems on certain types of hardware.  Reports came in on
the first release after this change.

On my own hardware, I have a laptop that sometimes stalls completely
before starting its display manager.  On a full power cycle it works,
but still very suspicious.

Lets just revert this change and wait for more feedback.  We can restore
it later if it's not the culprit.