Disable expert RCU configuration
Besides evidence of RCU misbehaving with MuQSS, the expert configuration
set in Liquorix is technically now the default if you enable expert RCU
settings.  RCU priority boosting is still off by default, but its
benefits are dubious.

Lets remove the expert RCU configuration so better allow the kernel to
properly set itself if it detects a more optimal configuration for the
hardware running.