config: Enable threaded preemptable tree RCU
This option was disabled before because it wasn't clear how much benefit
it was providing.  However, since Liquorix is re-focusing on low latency
with 1000hz tick rate, we want MuQSS to schedule the RCU threads.

Also, through some basic research online, it seems that setting the RCU
boost delay to 331ms is relatively ideal/safe.  331 is a prime number,
so it'll reduce the amount of contention that RCU would normally get at
the 500ms interval.  In these examples, they also recommend increasing
the priority, but MuQSS already does a much better job than CFS at
respecting priority levels, so lets not mess with that.