bin/gencontrol.py: Always set minor version to 0
After backporting updates from linux-siduction to liquorix, the
gencontrol.py script began including the minor version into the control
file correctly.  However, I didn't realize that this section of the code
is simply hard coded, and new updates quickly revealed this problem.

Set it back to zero until we come up with an automated way of populating
the kernel minor version in the package name.