config: Disable transcendent memory options
Disable both cleancache and frontswap.  Besides the fact that these
options help store more cache in memory, it seems to exaggerate the
effects of high memory pressure (on my systems).  This also means that
zswap won't be available, but maybe this is for the better considering
zram lets you stream using all threads available on your processor, and
thus, more performant.