config: Disable "conservative" acpi-cpufreq governor
This governor makes no sense on modern processors.  For the most part,
racing to finish (aggressive ondemand), or tuning the
up_threshold/no_turbo parameters will get you the frequency selection
behavior you're looking for.

Disable it so the chance of the wrong governor being selected is minimized
on Liquorix.