config: Synchronize EFI/I2C configuration from Arch Linux
Per report on techpatterns.com, fwupdmgr is unable to update firmware
with Liquorix [1].  The reporter identified some missing EFI
config options.

In order to synchronize these options, CONFIG_I2C needs to be set to
'y'.  With that, over 100 options became available and were set to 'm'
when possible, otherwise set to 'y'.

[1] https://techpatterns.com/forums/about2816.html