config: Reduce CONFIG_FBDEV_OVERALLOC from 300 to 100
As reported in #53, setting `rotate_all` for `fbcon` fails with Liquorix
but not other kernels.  Investigating configuration differences,
Liquorix uses a 300% overallocation for FBDEV to emulate triple
buffering.  However, in both Arch and Debian's latest configurations,
this value is set to 100.

Use the more common value so Liquorix behaves more similarly to common
stock kernels.