config: Relax control over ACPI resources
Relaxing control on ACPI resources seems bring back sensors for a lot of
modules on many of my systems.  Specifically this helps a lot on ASUS
motherboards as outlined in the bug below:

https://bugzilla.kernel.org/show_bug.cgi?id=204807

If this configuration reduces stability on a target system, one must
only pass `acpi_enforce_resources=strict` into their kernel boot
parameters and it will override this change.