debian/templates: Remove unnecessary/legacy package conflicts
A conflict for r8168-dkms has existed in the Liquorix templates since
the initial commit in 9974a81b14fc since May 2016.  Maybe at some point
Siduction was bundling its own version of the r8168 package.

Liquorix has never done that so we simply made it impossible to support
this hardware in Debian and Ubuntu.

Remove so this DKMS package can be used.

Fixes: https://github.com/damentz/liquorix-package/issues/94