config: Switch from ZSTD to XZ kernel image compression
Since we know that ZSTD causes strange issues on older distributions
Liquorix is built for, why not also make everything uniform and use XZ
compression everywhere.  The only impact will be a slower boot time
during decompression, but it's marginal and only obvious on systems so
slow that they're not supported by Liquorix anyways.