make-debian-release.sh: Always purge superseded PPA packages
Since this repository is for the most part unmonitored, when the repo
does get full, it silently ignores new submissions.  To fix that, always
purge superseded packages before submitting new ones so we're basically
"double buffering" with the last latest package and the one that's about
to be built and published.