Drop Ubuntu Focal, automate commits of changes made through build
Due to the lack of libtraceevent-dev in Ubuntu Focal, support is dropped
since without dependency, build fails when making perf.

Also, newer versions of dpkg-source / dpkg-buildpackage now stop the
build when changes are detected since the build started.  We can work
around this by always committing changes in a non-interactive fashion by
making `cat` the editor and preconfiguring the commit message and path
for `dpkg-source --commit`.

And last, switch to Debian Bookworm so we use its build tooling.