config: Switch acpi-cpufreq from module to built-in
Per recent issue [1], acpi-cpufreq appears to be missing from some users
systems, causing poor performance and confusing behavior.  Most likely,
the maintainer of the distribution or the user blacklisted the module on
purpose or by accident, causing Liquorix to misbehave.

To fix, compile acpi-cpufreq into the kernel so it can't be blacklisted.
This matches the behavior of amd-pstate and intel-pstate as well.

[1] https://github.com/damentz/liquorix-package/issues/161