Enable CONFIG_FTRACE_SYSCALLS
CONFIG_FTRACE_SYSCALLS is enabled by default in every distribution kernel
I've examined, Liquorix is the outlier.

Enabling this allows the tracing of filesystem syscalls. This allows for more
detailed event tracing with eBPF and 'perf'. These tracepoints are important
when trying to pinpoint performance issues with various applications. They
are also required for certain security software such as CloudStrike.