config: Disable CONFIG_ZBUD
After z3fold was deprecated, I looked for discussions about the zswap
pool allocators.  It turns out that the maintainers are actively getting
rid of all allocators except zsmalloc, and recommended for use with
ZSWAP.

The only reason zbud is still around is that it's been battle tested,
but it doesn't mean it's good.  It uses far more memory than zsmalloc
for a small advantage in latency.

In-case users have bad configurations not using zsmalloc, turn off zbud
so zswap always initializes with the correct allocator.

Links:
 - https://lore.kernel.org/lkml/CAJD7tkbRF6od-2x_L8-A1QL3=2Ww13sCj4S3i4bNndqF+3+_Vg@mail.gmail.com/
 - https://lore.kernel.org/linux-mm/CAKEwX=P21VvVyfmAADzXe0=Mqz3Htyx9nQuiivLchcDZRumh6Q@mail.gmail.com/