debian/templates: Define conflict with vhba-dkms
In #201, when vhba-dkms is installed with Liquorix, the package fails to
write the built vhba module because Liquorix already builds with it as
part of feature addition in the zen-sauce branch of Zen Kernel.

Set up a conflict with vhba-dkms for Debian and Ubuntu to prevent this
DKMS package to be installed concurrently with Liquorix.